move canvas to own file
